# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A Tag Service API in Go useing Gin Framework",
        "title": "Tag Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/api",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AuthLoginPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/dto.AuthLoginResponse"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Register",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register",
                "parameters": [
                    {
                        "description": "payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AuthRegisterPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/dto.AuthRegisterResponse"
                        }
                    }
                }
            }
        },
        "/auth/user": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get user by auth",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Get user by auth",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/dto.AuthUserResponse"
                        }
                    }
                }
            }
        },
        "/user": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get users",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "keyword",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "orderBy",
                        "name": "orderBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sortedBy",
                        "name": "sortedBy",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/dto.UserIndexResponse"
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/dto.UserShowResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.AuthLoginPayload": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "minLength": 3,
                    "example": ""
                },
                "username": {
                    "type": "string",
                    "minLength": 3,
                    "example": ""
                }
            }
        },
        "dto.AuthLoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": ""
                }
            }
        },
        "dto.AuthRegisterDataResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": ""
                },
                "id": {
                    "type": "string",
                    "example": ""
                },
                "name": {
                    "type": "string",
                    "example": ""
                },
                "updated_at": {
                    "type": "string",
                    "example": ""
                },
                "username": {
                    "type": "string",
                    "example": ""
                }
            }
        },
        "dto.AuthRegisterPayload": {
            "type": "object",
            "required": [
                "name",
                "password",
                "username"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "minLength": 3,
                    "example": ""
                },
                "password": {
                    "type": "string",
                    "minLength": 3,
                    "example": ""
                },
                "username": {
                    "type": "string",
                    "minLength": 3,
                    "example": ""
                }
            }
        },
        "dto.AuthRegisterResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.AuthRegisterDataResponse"
                },
                "message": {
                    "type": "string",
                    "example": ""
                }
            }
        },
        "dto.AuthUserDataResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": ""
                },
                "id": {
                    "type": "string",
                    "example": ""
                },
                "name": {
                    "type": "string",
                    "example": ""
                },
                "updated_at": {
                    "type": "string",
                    "example": ""
                },
                "username": {
                    "type": "string",
                    "example": ""
                }
            }
        },
        "dto.AuthUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.AuthUserDataResponse"
                },
                "message": {
                    "type": "string",
                    "example": ""
                }
            }
        },
        "dto.UserIndexDataResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": ""
                },
                "id": {
                    "type": "string",
                    "example": ""
                },
                "name": {
                    "type": "string",
                    "example": ""
                },
                "updated_at": {
                    "type": "string",
                    "example": ""
                },
                "username": {
                    "type": "string",
                    "example": ""
                }
            }
        },
        "dto.UserIndexResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.UserIndexDataResponse"
                    }
                },
                "message": {
                    "type": "string",
                    "example": ""
                },
                "pagination": {
                    "$ref": "#/definitions/helper.Paginate"
                }
            }
        },
        "dto.UserShowDataResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": ""
                },
                "id": {
                    "type": "string",
                    "example": ""
                },
                "name": {
                    "type": "string",
                    "example": ""
                },
                "updated_at": {
                    "type": "string",
                    "example": ""
                },
                "username": {
                    "type": "string",
                    "example": ""
                }
            }
        },
        "dto.UserShowResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.UserShowDataResponse"
                },
                "message": {
                    "type": "string",
                    "example": ""
                }
            }
        },
        "helper.Paginate": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 0
                },
                "page": {
                    "type": "integer",
                    "example": 0
                },
                "total": {
                    "type": "integer",
                    "example": 0
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}